[TOL] Functionalities moved to common parent
